personality_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.12.0
numpy==1.23.5
scikit-learn==1.3.2
python-multipart==0.0.6
pydantic==2.5.2
gunicorn==21.2.0
pandas==2.1.4